callflow_diagrams: allowed additional lines of text under drawn lines
There is new possible element nested in the "call" element - "text-line" - with
additional line of text added under the arrow. This element can repeat so more
lines of additional text are allowed (see callflow_diagrams/examples/example.xml)

During testing the "whitebox" under the diagram which was commented out is
added again, we will see if it shows to be handy.